refactor(models): restructure APIResponse and feature models
