+++ PingMe/Features/Auth/Models/AuthResponses.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+struct RegisterResponseData: Codable {
+    let email: String
+}
+
+struct VerifyResponseData: Codable {
+    let user: User
+    let tokens: Tokens
+}
