+++ PingMe/Shared/Models/User.swift
@@ -1,0 +1,26 @@
+import Foundation
+
+struct User: Codable {
+    let id: UUID
+    let email: String
+    let name: String
+    let username: String?
+    let phoneNumber: String?
+    let isOnline: Bool
+    let isVerified: Bool
+    let authProvider: String
+    let mailingMethod: String
+    let createdAt: Date
+    let updatedAt: Date
+
+    enum CodingKeys: String, CodingKey {
+        case id, email, name, username
+        case phoneNumber = "phone_number"
+        case isOnline = "is_online"
+        case isVerified = "is_verified"
+        case authProvider = "auth_provider"
+        case mailingMethod = "mailing_method"
+        case createdAt = "created_at"
+        case updatedAt = "updated_at"
+    }
+}
